only check for objects and not arrays
Before this change it returned "true" for both arrays and objects, now it only supports objects.

This is not 100% tested and needs to be tested with all kind of different data. I needed this change to be able to run `$auth.check([1,2,3,4])` to make the function return true if a user has one of these role_id's.